pyproject.toml
[project]
name = "diary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.8.5",
    "discord-py==2.3.2",
    "pygithub==1.59.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
]
